requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.4.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.3.0
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_55txi4fy1u/croot/bottleneck_1731058642212/work
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
catboost==1.2.8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
coverage==7.8.0
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
Cython==3.1.1
databricks-sdk==0.50.0
datasets==3.5.1
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker==7.1.0
dulwich==0.22.8
dynaconf==3.2.10
et_xmlfile==2.0.0
evidently==0.7.2
exceptiongroup==1.2.2
executing==2.2.0
Faker==37.1.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.3
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
formulaic==1.1.1
frozenlist==1.6.0
fsspec==2025.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
installer==0.7.0
interface-meta==1.3.0
ipython==8.36.0
ipywidgets==8.1.6
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
joblib>=1.0.1
jupyterlab_widgets==3.0.14
kaleido==0.2.1
keyring==25.6.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lifelines==0.30.0
lightgbm==4.6.0
lime==0.2.0.1
linearmodels==6.1
litestar==2.15.2
litestar-htmx==0.4.1
llvmlite==0.44.0
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib>=3.4.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mlxtend==0.23.4
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multipart==1.2.1
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numexpr @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b3kvvt6tc6/croot/numexpr_1730215947700/work
numpy>=1.21.0
onnx==1.17.0
onnxruntime==1.21.1
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas>=1.3.0
parso==0.8.4
patsy==1.0.1
pbs-installer==2025.4.9
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
plotly==5.24.1
pluggy==1.5.0
poetry==2.1.2
poetry-core==2.1.2
polyfactory==2.21.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyhdfe==0.2.0
pynndescent==0.5.13
pyparsing==3.2.3
pypdfium2==4.30.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_66ud1l42_h/croot/python-dateutil_1716495741162/work
python-docx==1.1.2
python-dotenv>=0.19.0
python-pptx>=0.6.21
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a4b76c83ik/croot/pytz_1713974318928/work
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-click==1.8.8
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn>=0.24.2
scipy==1.15.2
seaborn>=0.11.1
sentencepiece==0.2.0
setuptools-scm==8.3.1
shap==0.47.2
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
statsmodels>=0.12.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
tqdm>=4.61.2
traitlets==5.14.3
transformers==4.51.3
trove-classifiers==2025.4.11.15
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
ujson==5.10.0
umap-learn==0.5.7
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
xattr==1.1.4
xgboost>=1.4.2
xlrd==2.0.1
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
yfinance>=0.2.36
